pyproject.toml
[project]
name = "charge-cheapest"
version = "1.0.0"
description = "Tibber Cheapest Hours Battery Charging Blueprint"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "ISC" }
keywords = ["home-assistant", "blueprint", "tibber", "battery-charging"]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pyyaml>=6.0",
    "yamllint>=1.35",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 200
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
